Update messaging about pip for a more sophisticated audience
pip is a glorious library. I donate to it. It makes me happy. I am greatful to
all package managers, and they have a tough job. Rather than badmouthing pip, I'll
just give users an option for what to do on the offchance that it fails.

It probably will, and they'll have to edit paths and symlinks and stuff, but they will know that I never lost faith.

Also, the initial pip badmouthing came because it was for people who needed a quick and dirty way to make this run. For people who are more experienced, making pip work is probably worth it.

Review: The Internet