Re-enable support for time, and generate fake time data
While getting the guacamole library ready for wide release I accidentally broke
support for time in the model.  This commit restores that support.

Testing to ensure this actually increases the quality of learned models is still to be done.

Test Plan: Run new model and examine theta_flat. See that new values
exist in the places where they used to be arrays of zeros.

Auditors: jascha