DRY out normalizing time
Before, to get the log time taken we did a bunch of work to clean it up and make it
presentable. Now we just call a simple function that does that for us.

This is nice because it makes things simpler.

Test Plan: Run everything and it all passes. Jeez I sure should write some tests

Auditor: Jascha